+++ packages/twenty-front/src/modules/ui/navigation/step-bar/components/Step.tsx
@@ -3,7 +3,7 @@ import { motion } from 'framer-motion';
 import { useContext } from 'react';
 
 import { AnimatedCheckmark } from '@/ui/navigation/step-bar/components/internal/AnimatedCheckmark/AnimatedCheckmark';
-import { useIsMobile } from '@/ui/utilities/responsive/hooks/useIsMobile';
+import { useIsMobile } from 'twenty-ui/utilities';
 import {
   MOBILE_VIEWPORT,
   ThemeContext,