+++ packages/twenty-ui/src/utilities/index.ts
@@ -19,7 +19,6 @@ export { getOsShortcutSeparator } from './device/getOsShortcutSeparator';
 export { getUserDevice } from './device/getUserDevice';
 export { useIsMobile } from './responsive/hooks/useIsMobile';
 export { useIsTouchDevice } from './responsive/hooks/useIsTouchDevice';
-export { useScreenSize } from './screen-size/hooks/useScreenSize';
 export type { ClickOutsideAttributes } from './types/ClickOutsideAttributes';
 export type { Nullable } from './types/Nullable';
 export { getSafeUrl } from './utils/getSafeUrl';